LayoutTests/http/tests/history/back-with-fragment-change-expected.txt

main frame - has 1 onunload handler(s)
CONSOLE MESSAGE: Visited fragment and waited.
PASS

============== Back Forward List ==============
        http://127.0.0.1:8000/history/back-with-fragment-change.py
        http://127.0.0.1:8000/history/resources/back-with-fragment-change-target.html
curr->  http://127.0.0.1:8000/history/resources/back-with-fragment-change-target.html#foo
===============================================
